@@ -8,23 +8,30 @@
 	<packaging>war</packaging>
 	<properties>
 		<failOnMissingWebXml>false</failOnMissingWebXml>
+		<maven.compiler.source>11</maven.compiler.source>
+		<maven.compiler.target>11</maven.compiler.target>
 	</properties>
 	<dependencies>
 
 		<dependency>
 			<groupId>org.hibernate</groupId>
 			<artifactId>hibernate-entitymanager</artifactId>
-			<version>5.4.29.Final</version>
+			<version>5.4.30.Final</version>
 
 		</dependency>
 
-		<!--   <dependency>
+		<dependency>
+			<groupId>org.glassfish.jersey.core</groupId>
+			<artifactId>jersey-client</artifactId>
+			<version>2.31</version>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
 			<groupId>javax.servlet</groupId>
 			<artifactId>servlet-api</artifactId>
 			<version>2.5</version>
 			<scope>provided</scope>
-		</dependency>-->
-		
+		</dependency>
 		<dependency>
 			<groupId>javax</groupId>
 			<artifactId>javaee-api</artifactId>
@@ -41,7 +48,7 @@
 		<dependency>
 			<groupId>org.slf4j</groupId>
 			<artifactId>slf4j-log4j12</artifactId>
-			<version>1.7.28</version>
+			<version>1.7.5</version>
 		</dependency>
 
 		<dependency>
@@ -75,22 +82,45 @@
 			<version>2.3.8</version>
 		</dependency>
 
+		<dependency>
+			<groupId>io.jsonwebtoken</groupId>
+			<artifactId>jjwt</artifactId>
+			<version>0.9.1</version>
+		</dependency>
 
 	</dependencies>
 	<build>
 		<plugins>
-			<!-- <plugin> <groupId>io.openapitools.swagger</groupId> <artifactId>swagger-maven-plugin</artifactId> 
-				<configuration> <resourcePackages> <resourcePackage>io.openapitools.swagger.example</resourcePackage> 
-				<resourcePackage>io.openapitools.swagger.example.alternate</resourcePackage> 
-				</resourcePackages> <outputDirectory>${basedir}/target/</outputDirectory> 
-				<outputFilename>swagger</outputFilename> <outputFormats>JSON,YAML</outputFormats> 
-				<prettyPrint>true</prettyPrint> </configuration> <executions> <execution> 
-				<goals> <goal>generate</goal> </goals> </execution> </executions> </plugin> 
-				<plugin> <groupId>org.apache.maven.plugins</groupId> <artifactId>maven-compiler-plugin</artifactId> 
-				<version>3.8.1</version> <configuration> <source>1.8</source> <target>1.8</target> 
-				</configuration> </plugin> <plugin> <groupId>org.apache.maven.plugins</groupId> 
-				<artifactId>maven-javadoc-plugin</artifactId> <version>3.0.0</version> <configuration> 
-				<source>1.8</source> <target>1.8</target> </configuration> </plugin> -->
+			<plugin>
+				<groupId>io.openapitools.swagger</groupId>
+				<artifactId>swagger-maven-plugin</artifactId>
+				<version>2.1.5</version>
+				<configuration>
+					<resourcePackages>
+						<resourcePackage>com.carmen.app.boundary</resourcePackage>
+					</resourcePackages>
+					<outputDirectory>${basedir}/target/</outputDirectory>
+					<outputFilename>swagger</outputFilename>
+					<outputFormats>JSON,YAML</outputFormats>
+					<prettyPrint>true</prettyPrint>
+				</configuration>
+				<executions>
+					<execution>
+						<goals>
+							<goal>generate</goal>
+						</goals>
+					</execution>
+				</executions>
+			</plugin>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-war-plugin</artifactId>
+				<version>3.3.1</version>
+				<configuration>
+					 <!--  <outputDirectory>C:\ProgramFiles\payara5\glassfish\domains\domain1\autodeploy</outputDirectory>-->
+				</configuration>
+			</plugin>
+
 
 		</plugins>
 
