@@ -1,9 +1,6 @@
 package com.carmen.app.boundary;
 
-
-
 import java.util.List;
-//import java.util.UUID;
 import java.util.stream.Collectors;
 
 import javax.ejb.EJB;
@@ -23,35 +20,28 @@ import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 import javax.ws.rs.core.SecurityContext;
 
-//import org.apache.log4j.LogManager;
-//import org.apache.log4j.Logger;
+import com.carmen.app.control.CountryService;
+import com.carmen.app.dto.CountryDto;
 
-import com.carmen.app.control.CarService;
-import com.carmen.app.dto.CarDto;
-import com.carmen.app.entities.Car;
 import com.carmen.app.exceptions.EntityNotFoundException;
-
 import com.carmen.app.utils.Logged;
 import com.carmen.app.utils.Secured;
 
-
 /**
- * 
- * Resource that maps {@link Car} API to a method
  * 
  * @author Carmen Pi�era Moreno
  *
  */
-@Path("/cars")
+
+
+@Path("/countries")
 @Consumes(MediaType.APPLICATION_JSON)
 @Produces(MediaType.APPLICATION_JSON)
-
 @Logged
-public class CarResource implements ICarResource {
+public class CountryResource implements ICountryResource {
 
-	@EJB
-	private CarService carService;
-
+	@EJB 
+	private CountryService countryService;
 	
 	@Context
 	SecurityContext securityContext;
@@ -59,79 +49,70 @@ public class CarResource implements ICarResource {
 	
 	@GET
 	@Secured
-	
-	public Response getCars(@DefaultValue("0") @QueryParam(value = "page") int page,
+	public Response getCountries(@DefaultValue("0") @QueryParam(value = "page") int page,
 			@DefaultValue("20") @QueryParam(value = "size") int size,
 		 @QueryParam(value = "filterBy") String filterBy,
 			@QueryParam(value = "orderBy") String orderBy,
 			@QueryParam(value= "sort") @DefaultValue("asc") String sort) {
-		
 		if(securityContext.isUserInRole("username") || securityContext.isUserInRole("admin")) {
-			List<CarDto> cars = this.carService.getCars(page, size, filterBy, sort, orderBy).stream().map(car -> car.convertToDto()).collect(Collectors.toList());
+		List<CountryDto> countries = this.countryService.getCountries(page, size, filterBy, sort, orderBy).stream().map(country -> country.convertToDto()).collect(Collectors.toList());
 		
-		return Response.status(Status.OK).entity(cars).build();
-		}
+		return Response.status(Status.OK).entity(countries).build();
+		
+	}
 		else {
 			return Response.status(Status.UNAUTHORIZED).build();
 		}
-		
-	}
-
+		}
+	
+	
 	@GET
 	@Path("/{id}")
 	@Secured
-	public Response getCar(@PathParam("id") String id) {
+	public Response getCountry(@PathParam("id") String id) {
 		if(securityContext.isUserInRole("username") || securityContext.isUserInRole("admin")) {
 		try {
-			CarDto carDto = this.carService.getCar(id).convertToDto();
-			return Response.status(Status.OK).entity(carDto).build();
+			CountryDto countryDto = this.countryService.getCountry(id).convertToDto();
+			return Response.status(Status.OK).entity(countryDto).build();
 			
-
-		} catch (EntityNotFoundException ex) {
+		}catch (EntityNotFoundException ex) {
 			return Response.status(Status.NOT_FOUND).build();
-
-			
 		}
 		}else {
 			return Response.status(Status.UNAUTHORIZED).build();
 		}
-
 	}
+	
 
 	@PUT
 	@Path("/{id}")
 	@Secured
-	public Response updateCar(@PathParam("id") String id, CarDto carDto) {
+	public Response updateCountry(@PathParam("id") String id, CountryDto countryDto) {
 		if(securityContext.isUserInRole("admin")) {
-			
-		
 		try {
 
-			carDto.setId(id);
+			countryDto.setId(id);
 			
-			Response response = Response.status(Status.OK).entity(this.carService.updateCar(carDto.convertToEntity())).build();
-			return response;
+			return Response.status(Status.OK).entity(this.countryService.updateCountry(countryDto.convertToEntity())).build();
+			
 
 		} catch (EntityNotFoundException ex) {
-			Response response = Response.status(Status.NOT_FOUND).build();
+		return Response.status(Status.NOT_FOUND).build();
 
-			return response;
-}
+		}
+
 		}else {
 			return Response.status(Status.UNAUTHORIZED).build();
 		}
 
 	}
-
+	
 	@POST
 	@Secured
-	public Response createdCar(CarDto carDto) {
-
+	public Response createdCountry(CountryDto countryDto) {
 		if(securityContext.isUserInRole("admin")) {
-			
-		
-		this.carService.createCar(carDto.convertToEntity()).convertToDto();
-		Response response = Response.status(Status.CREATED).entity(carDto).build();
+		this.countryService.createCountry(countryDto.convertToEntity()).convertToDto();
+		Response response = Response.status(Status.CREATED).entity(countryDto).build();
 
 		return response;
 		}else {
@@ -140,20 +121,19 @@ public class CarResource implements ICarResource {
 	}
 
 	@DELETE
-	@Secured
 	@Path("/{id}")
 
-	public Response deleteCar(@PathParam("id") String id) {
+	public Response deleteCountry(@PathParam("id") String id) {
 		if(securityContext.isUserInRole("admin")) {
 		try {
-			this.carService.deleteCar(id);
+			this.countryService.deleteCountry(id);
 			return Response.status(Status.NO_CONTENT).build();
-			
+		
 		} catch (EntityNotFoundException ex) {
 
 			return Response.status(Status.NOT_FOUND).build();
-
 		}
+		
 		}else {
 			return Response.status(Status.UNAUTHORIZED).build();
 		}
@@ -161,5 +141,6 @@ public class CarResource implements ICarResource {
 	}
 
 	
-
+	
+	
 }
