@@ -6,18 +6,16 @@ import java.util.UUID;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-//import javax.persistence.GeneratedValue;
-//import javax.persistence.GenerationType;
+
 import javax.persistence.Id;
 import javax.persistence.NamedQuery;
-import javax.persistence.PrePersist;
+import javax.persistence.*;
 import javax.persistence.Table;
-//import javax.persistence.Temporal;
-//import javax.persistence.TemporalType;
+
 import javax.validation.constraints.NotNull;
 
 import org.hibernate.annotations.CreationTimestamp;
-//import org.hibernate.annotations.GenericGenerator;
+
 import org.hibernate.annotations.UpdateTimestamp;
 
 import com.carmen.app.dto.CarDto;
@@ -37,22 +35,22 @@ public class Car implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Id
-	//@GenericGenerator(name = "UUID", strategy = "uuid2")
-	//@GeneratedValue(generator = "UUID")
+	
 	private String id;
 
-	@Column(name = "brand_car")
+	
+	@ManyToOne
 	@NotNull(message = "Brand is required")
-	private String brand;
+	private Brand brand;
 
-	// @Temporal(TemporalType.TIMESTAMP)
-	@Column(name = "registration_date")
+	//@Temporal(TemporalType.TIMESTAMP)
+	@Column(name = "registration")
 	@NotNull(message = "Registration date is required")
 	private LocalDateTime registration;
 
-	@Column(name = "country")
+	@ManyToOne
 	@NotNull(message = "Country is required")
-	private String country;
+	private Country country;
 
 	// @Temporal(TemporalType.TIMESTAMP)
 	@Column(name = "created")
@@ -60,7 +58,7 @@ public class Car implements Serializable {
 	private LocalDateTime createdAt;
 
 	//@Temporal(TemporalType.TIMESTAMP)
-	@Column(name = "lastUpdate")
+	@Column(name = "lastupdate")
 	@UpdateTimestamp
 	private LocalDateTime lastUpdate;
 
@@ -68,7 +66,7 @@ public class Car implements Serializable {
 
 	}
 
-	public Car(String brand, LocalDateTime registration, String country) {
+	public Car(Brand brand, LocalDateTime registration, Country country) {
 	
 		this.brand = brand;
 		this.registration = registration;
@@ -84,11 +82,11 @@ public class Car implements Serializable {
 		this.id = id;
 	}
 
-	public String getBrand() {
+	public Brand getBrand() {
 		return brand;
 	}
 
-	public void setBrand(String brand) {
+	public void setBrand(Brand brand) {
 		this.brand = brand;
 	}
 
@@ -100,11 +98,11 @@ public class Car implements Serializable {
 		this.registration = registration;
 	}
 
-	public String getCountry() {
+	public Country getCountry() {
 		return country;
 	}
 
-	public void setCountry(String country) {
+	public void setCountry(Country country) {
 		this.country = country;
 	}
 
@@ -123,7 +121,6 @@ public class Car implements Serializable {
 	public void setLastUpdate(LocalDateTime lastUpdate) {
 		this.lastUpdate = lastUpdate;
 	}
-
 
 	@PrePersist
 	private void autoGeneratedId() {
